Completed get user location
